Implements a fairly OK api for dealing with users
This implements an API for dealing with users. It also renames the VolunteerProvider to UserProvider, since this is what the class does now.

The class provides an abstraction for thinking about users. It can be a wrapper around a database or could be used with other services such as google accounts, etc.