This adds a docker-compose file that can be used to run the project
This is meant to provide a basis for testing databases as well as to
be the prototype for an "official" environment.

This may also be used when we are setting up cloud deployments for
this system, although those will likely require their own non-docker
configurations. This may still be used as a reference to understand
the nessesary components in those cases.

If you are unfamiliar with docker, take a look at https://docs.docker.com/get-started/overview/

Basically, docker is a tool that lets you run a second operating system
inside your current operating system. This second operating system cannot
easily access the resources you have in your outer operating system - such
as most files or devices attached to your computer. However, the inner
operating system can run whatever programs you want, which makes it
easy to set up exactly what we want for cityzen server. This technique is
called "containerization." See https://en.wikipedia.org/wiki/OS-level_virtualization
for more information.