Adds a simple abstraction for sending messages to users
Adds MessageProvider, which can be used to represent a 3rd party service
which handles sending messages to users, such as an IMAP server or a
Twillio account.

Adds the POST /users/:id/report route. When a POST request is made
against this route, the app will attempt to send a message to the user
given by :id.

Also adds various minor improvements to other areas of the app to
implement to POST /users/:id/report route and tests.

TODO: make the contents of reports configurable